Switch to GLFW3,major network update(player can now move on server side too),camera controllers,player+grenade physics,+naive chunk renderin

+++ particle.h
@@ -2,8 +2,8 @@ typedef struct {
 	float x,y,z;
 	float vx,vy,vz;
 	float size;
-	long created;
-	long fade;
+	float created;
+	float fade;
 	unsigned int color;
 	boolean alive;
 } Particle;
